TOB-5283 Gjennomgang og navngiving
