+++ src/pages/Hovedside.module.css
@@ -1,0 +1,4 @@
+.page {
+    max-width: 768px;
+    margin: auto;
+}
